requirements.txt
Django==3.2.7
wheel==0.36.2
mysqlclient==2.0.3
djangorestframework-simplejwt==4.7.2
djangorestframework==3.12.4
PyJWT==2.1.0
django-phonenumber-field[phonenumbers]==5.2.0
django-cors-headers==3.8.0